+++ lib/src/reverse_engineering/pages/watch_page.dart
@@ -1,7 +1,8 @@
+import 'dart:convert';
+
 import 'package:collection/collection.dart';
 import 'package:html/dom.dart';
 import 'package:html/parser.dart' as parser;
-import 'package:meta/meta.dart';
 
 import '../../../youtube_explode_dart.dart';
 import '../../extensions/helpers_extension.dart';
@@ -13,31 +14,30 @@ import 'player_config_base.dart';
 
 ///
 class WatchPage extends YoutubePage<WatchPageInitialData> {
+  final VideoId videoId;
+
   static final RegExp _videoLikeExp =
       RegExp(r'"label"\s*:\s*"([\d,\.]+) likes"');
   static final RegExp _videoDislikeExp =
       RegExp(r'"label"\s*:\s*"([\d,\.]+) dislikes"');
-  static final RegExp _visitorInfoLiveExp =
-      RegExp('VISITOR_INFO1_LIVE=([^;]+)');
-  static final RegExp _yscExp = RegExp('YSC=([^;]+)');
 
   @override
   // Overridden to be non-nullable.
   // ignore: overridden_fields
   final Document root;
 
-  @internal
-  final String visitorInfoLive;
+  /// Cookies linked to this webpage
+  final Map<String, String> cookies;
 
-  @internal
-  final String ysc;
+  String get cookieString =>
+      cookies.entries.map((e) => '${e.key}=${e.value}').join('; ');
 
   ///
   String? get sourceUrl {
     final url = root
         .querySelectorAll('script')
         .map((e) => e.attributes['src'])
-        .whereNotNull()
+        .nonNulls
         .firstWhereOrNull((e) => e.contains('player_ias') && e.endsWith('.js'));
     if (url == null) {
       return null;
@@ -90,7 +90,15 @@ class WatchPage extends YoutubePage<WatchPageInitialData> {
 
   late final WatchPlayerConfig? playerConfig = getPlayerConfig();
 
-  late final PlayerResponse? playerResponse = getPlayerResponse();
+  late final PlayerResponse? playerResponse = getInitialPlayerResponse();
+
+  late final Map<String, dynamic> ytCfg = _getYtCfg();
+
+  Map<String, dynamic> _getYtCfg() {
+    return json.decode(RegExp(r'ytcfg\.set\s*\(\s*({.+?})\s*\)\s*;')
+        .firstMatch(root.outerHtml)!
+        .group(1)!);
+  }
 
   ///
   WatchPlayerConfig? getPlayerConfig() {
@@ -104,7 +112,7 @@ class WatchPage extends YoutubePage<WatchPageInitialData> {
     return WatchPlayerConfig(jsonMap);
   }
 
-  PlayerResponse? getPlayerResponse() {
+  PlayerResponse? getInitialPlayerResponse() {
     final scriptText = root
         .querySelectorAll('script')
         .map((e) => e.text)
@@ -120,22 +128,38 @@ class WatchPage extends YoutubePage<WatchPageInitialData> {
   }
 
   ///
-  WatchPage.parse(String raw, this.visitorInfoLive, this.ysc)
+  WatchPage.parse(String raw, this.videoId, this.cookies)
       : root = parser.parse(raw),
         super(parser.parse(raw), (root) => WatchPageInitialData(root));
 
   ///
   static Future<WatchPage> get(YoutubeHttpClient httpClient, String videoId) {
     final url = Uri.parse(
-      'https://youtube.com/watch?v=$videoId&bpctr=9999999999&hl=en',
+      'https://www.youtube.com/watch?v=$videoId&bpctr=9999999999&has_verified=1&hl=en',
     );
-    return retry(httpClient, () async {
-      final req = await httpClient.get(url, validate: true);
+    const defaultCookies = 'PREF=hl=en&tz=UTC; SOCS=CAI; GPS=1';
+    const headers = {
+      'User-Agent':
+          'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/96.0.4664.18 Safari/537.36',
+      'Accept':
+          'text/html,application/xhtml+xml,application/xml;q=0.9,*/*;q=0.8',
+      'Accept-Language': 'en-us,en;q=0.5',
+      'Sec-Fetch-Mode': 'navigate',
+      'Cookie': defaultCookies,
+    };
 
-      final cookies = req.headers['set-cookie']!;
-      final visitorInfoLive = _visitorInfoLiveExp.firstMatch(cookies)?.group(1);
-      final ysc = _yscExp.firstMatch(cookies)!.group(1)!;
-      final result = WatchPage.parse(req.body, visitorInfoLive ?? '', ysc);
+    final cookiesExp = RegExp(r'(?:^|,)(\w.+?)=(.*?);');
+
+    return retry(httpClient, () async {
+      final req = await httpClient.get(url, headers: headers, validate: true);
+
+      final cookieHeader = req.headers['set-cookie']!;
+      final matches = cookiesExp.allMatches(cookieHeader);
+      final cookies = Map.fromEntries(
+          matches.map((e) => MapEntry(e.group(1)!, e.group(2)!)))
+        ..addAll({'PREF': 'hl=en', 'SOCS': 'CAI', 'GPS': '1'});
+
+      final result = WatchPage.parse(req.body, VideoId(videoId), cookies);
 
       if (!result.isOk) {
         throw TransientFailureException('Video watch page is broken.');