+++ lib/src/videos/streams/stream_client.dart
@@ -1,73 +1,112 @@
 import 'dart:collection';
 
+import 'package:logging/logging.dart';
+
 import '../../exceptions/exceptions.dart';
 import '../../extensions/helpers_extension.dart';
 import '../../retry.dart';
-import '../../reverse_engineering/cipher/cipher_manifest.dart';
 import '../../reverse_engineering/heuristics.dart';
 import '../../reverse_engineering/models/stream_info_provider.dart';
 import '../../reverse_engineering/pages/watch_page.dart';
 import '../../reverse_engineering/youtube_http_client.dart';
-import '../video_controller.dart';
 import '../video_id.dart';
+import '../youtube_api_client.dart';
 import 'stream_controller.dart';
 import 'streams.dart';
 
 /// Queries related to media streams of YouTube videos.
 class StreamClient {
+  final _logger = Logger('YoutubeExplode.StreamsClient');
   final YoutubeHttpClient _httpClient;
   final StreamController _controller;
 
   /// Initializes an instance of [StreamClient]
   StreamClient(this._httpClient) : _controller = StreamController(_httpClient);
 
-  CipherManifest? _cipherManifest;
-
   /// Gets the manifest that contains information
   /// about available streams in the specified video.
   ///
-  /// If [fullManifest] is set to `true`, more streams types will be fetched
-  /// and track of different languages (if present) will be included.
+  /// See [YoutubeApiClient] for all the possible clients that can be set using the [ytClients] parameter.
+  /// If [ytClients] is null the library automatically manages the clients, otherwise only the clients provided are used.
+  /// Currently by default the ios and android clients are used, if the extraction fails the tvSimply client is used instead (as of `v2.3.0 not yet supported due to signature deciphering not implemented).
+  ///
+  /// If the extraction fails an exception is thrown, to diagnose the issue enable the logging from the `logging` package, and open an issue with the output.
+  /// For example:
+  /// ```dart
+  /// Logger.root.level = Level.FINER;
+  /// Logger.root.onRecord.listen(print);
+  /// // run yt related code ...
+  ///
+  /// ```
   Future<StreamManifest> getManifest(dynamic videoId,
-      {bool fullManifest = false}) {
+      {@Deprecated(
+          'Use the ytClient parameter instead passing the proper [YoutubeApiClient]s')
+      bool fullManifest = false,
+      List<YoutubeApiClient>? ytClients}) async {
     videoId = VideoId.fromString(videoId);
+    final clients =
+        ytClients ?? [YoutubeApiClient.ios, YoutubeApiClient.android];
 
-    return retry(_httpClient, () async {
-      final streams =
-          await _getStreams(videoId, fullManifest: fullManifest).toList();
-      if (streams.isEmpty) {
-        throw VideoUnavailableException(
-          'Video "$videoId" does not contain any playable streams.',
-        );
-      }
+    final uniqueStreams = LinkedHashSet<StreamInfo>(
+      equals: (a, b) {
+        if (a.runtimeType != b.runtimeType) return false;
+        if (a is AudioStreamInfo && b is AudioStreamInfo) {
+          return a.tag == b.tag && a.audioTrack == b.audioTrack;
+        }
+        return a.tag == b.tag;
+      },
+      hashCode: (e) {
+        if (e is AudioStreamInfo) {
+          return e.tag.hashCode ^ e.audioTrack.hashCode;
+        }
+        return e.tag.hashCode;
+      },
+    );
 
-      final response = await _httpClient.head(streams.first.url);
-      if (response.statusCode == 403) {
-        throw YoutubeExplodeException(
-          'Video $videoId returned 403 (stream: ${streams.first.tag}',
-        );
-      }
+    Object? lastException;
 
-      // Remove duplicate streams (must have same tag and audioTrack (if it's an audio stream))
-      final uniqueStreams = LinkedHashSet<StreamInfo>(
-        equals: (a, b) {
-          if (a.runtimeType != b.runtimeType) return false;
-          if (a is AudioStreamInfo && b is AudioStreamInfo) {
-            return a.tag == b.tag && a.audioTrack == b.audioTrack;
+    for (final client in clients) {
+      _logger.fine(
+          'Getting stream manifest for video $videoId with client: ${client.payload['context']['client']['clientName']}');
+      try {
+        await retry(_httpClient, () async {
+          final streams = await _getStreams(videoId, ytClient: client).toList();
+          if (streams.isEmpty) {
+            throw VideoUnavailableException(
+              'Video "$videoId" does not contain any playable streams.',
+            );
           }
-          return a.tag == b.tag;
-        },
-        hashCode: (e) {
-          if (e is AudioStreamInfo) {
-            return e.tag.hashCode ^ e.audioTrack.hashCode;
-          }
-          return e.tag.hashCode;
-        },
-      );
-      uniqueStreams.addAll(streams);
 
-      return StreamManifest(uniqueStreams.toList());
-    });
+          final response = await _httpClient.head(streams.first.url);
+          if (response.statusCode == 403) {
+            throw YoutubeExplodeException(
+              'Video $videoId returned 403 (stream: ${streams.first.tag})',
+            );
+          }
+          uniqueStreams.addAll(streams);
+        });
+      } catch (e, s) {
+        _logger.severe(
+            'Failed to get stream manifest for video $videoId with client: ${client.payload['context']['client']['clientName']}. Reason: $e\n',
+            e,
+            s);
+        lastException = e;
+      }
+    }
+
+    // If the user has not provided any client retry with the tvSimply client.
+    // TODO: Enable when signature deciphering is implemented
+    /*
+        if (uniqueStreams.isEmpty && ytClients == null) {
+              return getManifest(videoId,
+          ytClients: [YoutubeApiClient.tvSimplyEmbedded]);
+    }*/
+    if (uniqueStreams.isEmpty) {
+      throw lastException ??
+          VideoUnavailableException(
+              'Video "$videoId" has no available streams');
+    }
+    return StreamManifest(uniqueStreams.toList());
   }
 
   /// Gets the HTTP Live Stream (HLS) manifest URL
@@ -101,47 +140,20 @@ class StreamClient {
   Stream<List<int>> get(StreamInfo streamInfo) =>
       _httpClient.getStream(streamInfo, streamClient: this);
 
-  Future<CipherManifest> _getCipherManifest() async {
-    if (_cipherManifest != null) {
-      return _cipherManifest!;
-    }
-
-    final playerSource = await _controller.getPlayerSource();
-    final manifest = playerSource.cipherManifest;
-    if (manifest == null) {
-      throw YoutubeExplodeException(
-        'Could not find the cipher operations manifest.',
-      );
-    }
-    return _cipherManifest = manifest;
-  }
-
   Stream<StreamInfo> _getStreams(VideoId videoId,
-      {required bool fullManifest}) async* {
-    try {
-      // Use await for instead of yield* to catch exceptions
-      await for (final stream
-          in _getStream(videoId, VideoController.androidTestSuiteClient)) {
-        yield stream;
-      }
-      if (fullManifest) {
-        await for (final stream
-            in _getStream(videoId, VideoController.androidClient)) {
-          yield stream;
-        }
-      }
-    } on VideoUnplayableException catch (e) {
-      if (e is VideoRequiresPurchaseException) {
-        rethrow;
-      }
-      yield* _getCipherStream(videoId);
+      {required YoutubeApiClient ytClient}) async* {
+    // Use await for instead of yield* to catch exceptions
+    await for (final stream in _getStream(videoId, ytClient)) {
+      yield stream;
     }
   }
 
-  Stream<StreamInfo> _getStream(VideoId videoId,
-      Map<String, Map<String, Map<String, Object>>> ytClient) async* {
+  Stream<StreamInfo> _getStream(
+      VideoId videoId, YoutubeApiClient ytClient) async* {
+    final watchPage = await WatchPage.get(_httpClient, videoId.value);
     final playerResponse =
-        await _controller.getPlayerResponse(videoId, ytClient);
+        await _controller.getPlayerResponse(videoId, ytClient, watchPage);
+
     if (!playerResponse.previewVideoId.isNullOrWhiteSpace) {
       throw VideoRequiresPurchaseException.preview(
         videoId,
@@ -159,44 +171,20 @@ class StreamClient {
         reason: playerResponse.videoPlayabilityError ?? '',
       );
     }
-    yield* _parseStreamInfo(playerResponse.streams, videoId);
+    yield* _parseStreamInfo(playerResponse.streams, watchPage);
 
     if (!playerResponse.dashManifestUrl.isNullOrWhiteSpace) {
       final dashManifest =
           await _controller.getDashManifest(playerResponse.dashManifestUrl!);
-      yield* _parseStreamInfo(dashManifest.streams, videoId);
+      yield* _parseStreamInfo(dashManifest.streams, watchPage);
     }
-  }
-
-  Stream<StreamInfo> _getCipherStream(VideoId videoId) async* {
-    final cipherManifest = await _getCipherManifest();
-    final playerResponse = await _controller.getPlayerResponseWithSignature(
-      videoId,
-      cipherManifest.signatureTimestamp,
-    );
-
-    if (!playerResponse.isVideoPlayable) {
-      throw VideoUnplayableException.unplayable(
-        videoId,
-        reason: playerResponse.videoPlayabilityError ?? '',
-      );
-    }
-
-    yield* _parseStreamInfo(playerResponse.streams, videoId);
   }
 
   Stream<StreamInfo> _parseStreamInfo(
-      Iterable<StreamInfoProvider> streams, VideoId videoId) async* {
+      Iterable<StreamInfoProvider> streams, WatchPage watchPage) async* {
     for (final stream in streams) {
       final itag = stream.tag;
       var url = Uri.parse(stream.url);
-      if (!stream.signature.isNullOrWhiteSpace) {
-        final cipherManifest = await _getCipherManifest();
-        url = url.setQueryParam(
-          stream.signatureParameter!,
-          cipherManifest.decipher(stream.signature!),
-        );
-      }
 
       final contentLength = stream.contentLength ??
           (await _httpClient.getContentLength(url, validate: false)) ??
@@ -230,7 +218,7 @@ class StreamClient {
             stream.source != StreamSource.adaptive) {
           assert(stream.audioTrack == null);
           yield MuxedStreamInfo(
-            videoId,
+            watchPage.videoId,
             itag,
             url,
             container,
@@ -249,7 +237,7 @@ class StreamClient {
 
         // Video only
         yield VideoOnlyStreamInfo(
-          videoId,
+          watchPage.videoId,
           itag,
           url,
           container,
@@ -267,7 +255,7 @@ class StreamClient {
         // Audio-only
       } else if (!audioCodec.isNullOrWhiteSpace) {
         yield AudioOnlyStreamInfo(
-            videoId,
+            watchPage.videoId,
             itag,
             url,
             container,